+++ caffe2/CMakeLists.txt
@@ -94,6 +94,7 @@ if(INTERN_BUILD_ATEN_OPS)
   list(APPEND Caffe2_DEPENDENCY_INCLUDE ${ATen_THIRD_PARTY_INCLUDE})
 endif()
 
+
 # ---[ Caffe2 build
 # Note: the folders that are being commented out have not been properly
 # addressed yet.
@@ -287,7 +288,19 @@ if(NOT TORCH_INSTALL_LIB_DIR)
   set(TORCH_INSTALL_LIB_DIR lib)
 endif()
 
+message(WARNING "Caffe2_CPU_INCLUDE before: ${Caffe2_CPU_INCLUDE}")
 
+if(NOT INTERN_BUILD_MOBILE OR NOT BUILD_CAFFE2_MOBILE)
+  if(USE_DISTRIBUTED)
+    add_subdirectory(${TORCH_SRC_DIR}/lib/c10d lib_c10d)
+    list(APPEND Caffe2_CPU_SRCS ${C10D_CPU_SRCS})
+    list(APPEND Caffe2_GPU_SRCS ${C10D_CUDA_SRCS})
+    list(APPEND Caffe2_CPU_INCLUDE ${C10D_INCLUDE})
+    list(APPEND Caffe2_PUBLIC_DEPENDENCY_LIBS ${C10D_LIBS})
+ endif()
+endif()
+
+message(WARNING "Caffe2_CPU_INCLUDE after: ${Caffe2_CPU_INCLUDE}")
 
 if(NOT INTERN_BUILD_MOBILE OR NOT BUILD_CAFFE2_MOBILE)
   if(USE_DISTRIBUTED)
@@ -298,8 +311,8 @@ if(NOT INTERN_BUILD_MOBILE OR NOT BUILD_CAFFE2_MOBILE)
     # up being empty. Downstream targets should also add a #ifdef guard.
     if(NOT WIN32)
       add_library(process_group_agent "${TORCH_SRC_DIR}/csrc/distributed/rpc/process_group_agent.cpp" "${TORCH_SRC_DIR}/csrc/distributed/rpc/process_group_agent.h")
-      target_link_libraries(process_group_agent PRIVATE torch c10d fmt::fmt-header-only)
-      add_dependencies(process_group_agent torch c10d)
+      target_link_libraries(process_group_agent PRIVATE torch fmt::fmt-header-only)
+      add_dependencies(process_group_agent torch)
 
       add_library(tensorpipe_agent
         "${TORCH_SRC_DIR}/csrc/distributed/rpc/tensorpipe_agent.cpp"
@@ -307,8 +320,8 @@ if(NOT INTERN_BUILD_MOBILE OR NOT BUILD_CAFFE2_MOBILE)
         "${TORCH_SRC_DIR}/csrc/distributed/rpc/tensorpipe_utils.cpp"
         "${TORCH_SRC_DIR}/csrc/distributed/rpc/tensorpipe_utils.h"
         )
-      target_link_libraries(tensorpipe_agent PRIVATE torch c10d tensorpipe fmt::fmt-header-only)
-      add_dependencies(tensorpipe_agent torch c10d)
+      target_link_libraries(tensorpipe_agent PRIVATE torch tensorpipe fmt::fmt-header-only)
+      add_dependencies(tensorpipe_agent torch )
       if(USE_TENSORPIPE)
         target_compile_definitions(tensorpipe_agent PUBLIC USE_TENSORPIPE)
         target_link_libraries(tensorpipe_agent PRIVATE tensorpipe)
@@ -897,10 +910,6 @@ endif()
     ${PROJECT_BINARY_DIR}/TorchConfig.cmake
     DESTINATION share/cmake/Torch)
 
-  if(USE_DISTRIBUTED)
-    add_subdirectory(${TORCH_SRC_DIR}/lib/c10d lib_c10d)
-  endif()
-
 
   # ---[ Torch python bindings build
   add_subdirectory(../torch torch)
@@ -967,12 +976,24 @@ if(USE_ROCM)
     )
 endif()
 
-# Pass USE_DISTRIBUTED to torch_cpu, as some codes in jit/pickler.cpp and
-# jit/unpickler.cpp need to be compiled only when USE_DISTRIBUTED is set
 if(USE_DISTRIBUTED)
+  # Pass USE_DISTRIBUTED to torch_cpu, as some codes in jit/pickler.cpp and
+  # jit/unpickler.cpp need to be compiled only when USE_DISTRIBUTED is set
   target_compile_definitions(torch_cpu PRIVATE
     USE_DISTRIBUTED
   )
+  if(USE_GLOO)
+    target_compile_definitions(torch_cpu INTERFACE USE_C10D_GLOO)
+  endif()
+  if(USE_NCCL)
+    target_compile_definitions(torch_cpu INTERFACE USE_C10D_NCCL)
+  endif()
+  if(USE_MPI)
+    target_compile_definitions(torch_cpu INTERFACE USE_C10D_MPI)
+  endif()
+
+  target_include_directories(torch_cpu PUBLIC $<BUILD_INTERFACE:${C10D_INCLUDE}>)
+ 
   # Pass USE_RPC in order to reduce use of
   # #if defined(USE_DISTRIBUTED) && !defined(_WIN32)
   # need to be removed when RPC is supported
