+++ torch/CMakeLists.txt
@@ -179,7 +179,7 @@ if(USE_DISTRIBUTED)
       list(APPEND TORCH_PYTHON_LINK_LIBRARIES tensorpipe)
       list(APPEND TORCH_PYTHON_PUBLIC_COMPILE_DEFINITIONS USE_TENSORPIPE)
     endif()
-    list(APPEND TORCH_PYTHON_LINK_LIBRARIES c10d)
+    #list(APPEND TORCH_PYTHON_LINK_LIBRARIES c10d)
     list(APPEND TORCH_PYTHON_COMPILE_DEFINITIONS USE_C10D)
 endif()
