+++ torch/lib/CMakeLists.txt
@@ -1,0 +1,10 @@
+# Export c10d cpu and cuda source file lists to parent scopes
+set(C10D_CPU_SRCS ${C10D_CPU_SRCS} PARENT_SCOPE)
+set(C10D_CPU_HDRS ${C10D_CPU_HDRS} PARENT_SCOPE)
+set(C10D_CUDA_SRCS ${C10D_CUDA_SRCS} PARENT_SCOPE)
+set(C10D_CUDA_HDRS ${C10D_CUDA_HDRS} PARENT_SCOPE)
+set(C10D_PUBLIC_HDRS ${C10D_PUBLIC_HDRS} PARENT_SCOPE) 
+set(C10D_INCLUDE ${C10D_INCLUDE} PARENT_SCOPE) 
+set(C10D_LIBS ${C10D_LIBS} PARENT_SCOPE) 
+
+message(STATUS "C10D_CPU_SRCS: ${C10D_CPU_SRCS}")
