+++ torch/lib/c10d/default_comm_hooks.cpp
@@ -1,4 +1,4 @@
-#include <torch/csrc/distributed/c10d/default_comm_hooks.h>
+#include <c10d/default_comm_hooks.hpp>
 
 #include <c10d/ProcessGroup.hpp>
 #include <torch/csrc/distributed/c10d/comm.h>
