+++ torch/lib/c10d/example/CMakeLists.txt
@@ -1,3 +1,3 @@
 add_executable(allreduce allreduce.cpp)
 target_include_directories(allreduce PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/..)
-target_link_libraries(allreduce pthread c10d)
+target_link_libraries(allreduce pthread torch)
