+++ torch/lib/c10d/test/ProcessGroupNCCLTest.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 
 #include <c10d/FileStore.hpp>
-#include <c10d/ProcessGroupNCCL.hpp>
+#include <c10d/cuda/ProcessGroupNCCL.hpp>
 #include <c10d/test/CUDATest.hpp>
 #include <c10d/test/TestUtils.hpp>
